new changes

+++ src/main/java/org/yearup/controllers/CategoriesController.java
@@ -1,9 +1,11 @@
 package org.yearup.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.security.access.annotation.Secured;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 import org.yearup.data.CategoryDao;
 import org.yearup.data.ProductDao;
 import org.yearup.models.Category;
@@ -11,22 +13,18 @@ import org.yearup.models.Product;
 
 import java.util.List;
 
-// add the annotations to make this a REST controller
-// add the annotation to make this controller the endpoint for the following url
-    // http://localhost:8080/categories
-// add annotation to allow cross site origin requests
 @CrossOrigin
 @RestController
-@RequestMapping("/categories")
+@RequestMapping("categories")
 public class CategoriesController
 
 {
-    @Autowired
-    private final CategoryDao categoryDao;
+
+    private CategoryDao categoryDao;
 
 
-    @Autowired
-    private final ProductDao productDao;
+
+    private  ProductDao productDao;
 
     @Autowired
     public CategoriesController(CategoryDao categoryDao, ProductDao productDao) {
@@ -34,22 +32,21 @@ public class CategoriesController
         this.productDao = productDao;
     }
 
-    public CategoryDao getCategoryDao() {
-        return categoryDao;
-    }
-
-    public ProductDao getProductDao() {
-        return productDao;
-    }
 // create an Autowired controller to inject the categoryDao and ProductDao
 
-    // add the appropriate annotation for a get action
+
+    @GetMapping("")
     public List<Category> getAll(){
+        try
+        {
+            return categoryDao.getAllCategories();
+        }
+        catch(Exception ex)
+        {
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Oops... our bad.");
+        }
 
-    List<Category> categories = categoryDao.getAllCategories();
 
-        // find and return all categories
-        return categories;
     }
 
     // add the appropriate annotation for a get action
@@ -59,8 +56,24 @@ public class CategoriesController
     @RequestMapping(path = "/categories/{id}")
     public Category getById(@PathVariable int id)
     {
+        {
+            try
+            {
+                var category = categoryDao.getById(id);
+
+                if(category == null)
+                    throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+
+                return category;
+            }
+            catch(Exception ex)
+            {
+                throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Oops... our bad.");
+            }
+        }
+
         // get the category by id
-        return categoryDao.getById(id);
+
     }
 
     // the url to return all products in category 1 would look like this
@@ -74,13 +87,21 @@ public class CategoriesController
 
     // add annotation to call this method for a POST action
     // add annotation to ensure that only an ADMIN can call this function
-    @PostMapping("/category")
-    @Secured("ROLE_ADMIN")
-    public Category addCategory(@RequestBody Category category)
-    {
-        // insert the category
-        return categoryDao.insert(category);
+    @PostMapping("")
+    @PreAuthorize("hasRole('ROLE_ADMIN')")
+    public Category addCategory(@RequestBody Category category) {
+
+        try {
+
+            return categoryDao.create(category);
+
+        } catch (Exception ex) {
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Oops... our bad.");
+        }
     }
+        // insert the category
+        //return categoryDao.insert(category);
+
 
     // add annotation to call this method for a PUT (update) action - the url path must include the categoryId
     // add annotation to ensure that only an ADMIN can call this function
@@ -99,7 +120,23 @@ public class CategoriesController
     @DeleteMapping("/{id}")
     @PreAuthorize("('ADMIN')")
     public void deleteCategory(@PathVariable int id) {
-        categoryDao.delete(id);
-        // delete the category by id
+        try
+        {
+            var category = categoryDao.getById(id);
+
+            if(category == null)
+                throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+
+            categoryDao.delete(id);
+        }
+        catch(Exception ex)
+        {
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Oops... our bad.");
+        }
     }
 }
+
+
+        // delete the category by id
+
+
